refactor: created custom widgets for landing page
