+++ lib/pages/landingPage/components/android_landing_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class AndroidLandingPage extends StatelessWidget {
+  const AndroidLandingPage({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
