+++ lib/pages/landingPage/components/landing_page_widgets.dart
@@ -1,0 +1,4 @@
+export 'package:ax_dapp/pages/landingPage/components/android_landing_page.dart';
+export 'package:ax_dapp/pages/landingPage/components/athletex_logo.dart';
+export 'package:ax_dapp/pages/landingPage/components/desktop_landing_page.dart';
+export 'package:ax_dapp/pages/landingPage/components/start_trading_button.dart';
