+++ lib/pages/landingPage/components/start_trading_button.dart
@@ -1,0 +1,55 @@
+import 'package:ax_dapp/pages/V1App.dart';
+import 'package:flutter/material.dart';
+
+class StartTradingButton extends StatelessWidget {
+  const StartTradingButton({
+    Key? key,
+    required this.isWeb,
+    required this.tradingTextSize,
+  }) : super(key: key);
+
+  final bool isWeb;
+  final double tradingTextSize;
+
+  @override
+  Widget build(BuildContext context) {
+    return isWeb
+      ? TextButton(
+        style: ButtonStyle(
+          shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+            RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(100),
+              side: BorderSide(color: Colors.amber[400]!)
+            )
+          )
+        ),
+        onPressed: () {
+          Navigator.push(
+              context, MaterialPageRoute(builder: (context) => V1App()));
+        },
+        child: Text(
+          "Start Trading",
+          style: TextStyle(color: Colors.amber[400]!, fontSize: tradingTextSize, fontFamily: 'OpenSans', fontWeight: FontWeight.w400),
+        ),
+      )         
+      : TextButton(
+        style: ButtonStyle(
+          backgroundColor: MaterialStateProperty.all<Color>(Colors.amber[300]!.withOpacity(0.15)),
+          shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+            RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(20),
+              side: BorderSide(color: Colors.transparent)
+            )
+          )
+        ),
+        onPressed: () {
+          Navigator.push(
+              context, MaterialPageRoute(builder: (context) => V1App()));
+        },
+        child: Text(
+          "Start",
+          style: TextStyle(color: Colors.amber[400]!, fontSize: tradingTextSize, fontFamily: 'OpenSans', fontWeight: FontWeight.w400),
+        ),
+      );      
+  }
+}
